+++ src/RtpVideoQueue.h
@@ -7,7 +7,7 @@ typedef struct _RTPV_QUEUE_ENTRY {
     struct _RTPV_QUEUE_ENTRY* prev;
     PRTP_PACKET packet;
     uint64_t receiveTimeUs;
-    uint32_t presentationTimeMs;
+    uint64_t presentationTimeMs;
     int length;
     bool isParity;
 } RTPV_QUEUE_ENTRY, *PRTPV_QUEUE_ENTRY;
@@ -43,7 +43,7 @@ typedef struct _RTP_VIDEO_QUEUE {
     uint8_t multiFecCurrentBlockNumber;
     uint8_t multiFecLastBlockNumber;
 
-    uint32_t lastOosFramePresentationTimestamp;
+    uint64_t lastOosFramePresentationTimestamp;
     bool receivedOosData;
 
     RTP_VIDEO_STATS stats; // the above values are short-lived, this tracks stats for the life of the queue
